backend/pyproject.toml
[project]
name = "realrent-backend"
version = "0.1.0"
description = "RealRent FastAPI backend"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "alembic>=1.14",
  "psycopg[binary]>=3.2",
  "pydantic-settings>=2.5",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["alembic*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
